add validators

+++ src/main/java/com/tiendavirtual/data/controllers/ProductoController.java
@@ -30,58 +30,74 @@ public class ProductoController {
 	
 	@RequestMapping("")
 	public String indexUsuario(@ModelAttribute("producto") Producto producto,Model model , HttpSession session) {
-		
 		if(session.getAttribute("userId") != null) {
 			List<Producto> lista_productos = pservice.findAll();
 			model.addAttribute("lista_productos", lista_productos);
-			
 			return "producto.jsp";
 		}else {
 			return "redirect:/login";
 		}
-		
-		
 	}
 	
 
 
 	@RequestMapping(value="/crear", method = RequestMethod.POST)
-	public String crearUsuario(@Valid @ModelAttribute("producto") Producto producto, Model model) {
-		if(producto.getNombre().isBlank() || producto.getPrecio().isNaN() || producto.getCaracteristica().isBlank() || producto.getStock()== null) {
-			model.addAttribute("error", "Todos los campos son requeridos!");
-			List<Producto> lista_producto = pservice.findAll();
-			model.addAttribute("lista_productos", lista_producto);
-			return "producto.jsp";
+	public String crearProducto(@Valid @ModelAttribute("producto") Producto producto, Model model, HttpSession session) {
+		
+		if(session.getAttribute("userId") != null) {
+			if(producto.getNombre().isBlank() || producto.getPrecio().isNaN() || producto.getCaracteristica().isBlank() || producto.getStock()== null) {
+				model.addAttribute("error", "Todos los campos son requeridos!");
+				List<Producto> lista_producto = pservice.findAll();
+				model.addAttribute("lista_productos", lista_producto);
+				return "producto.jsp";
+			}
+			Producto produ = pservice.insertarProducto(producto);
+			return "redirect:/producto";
+			
+		}else {
+			return "redirect:/login";
 		}
 		
 		
-		Producto produ = pservice.insertarProducto(producto);
-		return "redirect:/producto";
+		
+		
+		
 	}
 	
 	@RequestMapping(value="/actualizar/{id}", method = RequestMethod.GET)
-	public String actualizarProducto(@PathVariable("id") Long id, Model model) {
+	public String actualizarProducto(@PathVariable("id") Long id, Model model, HttpSession session) {
+		if(session.getAttribute("userId") != null) {
 		Producto producto = pservice.buscarProducto(id);
-		
 		model.addAttribute("producto", producto);
 		return "editar_producto.jsp";
+		}
+		else {
+			return "redirect:/login";
+		}
 	}
 	
 	@RequestMapping(value="/modificar",method= RequestMethod.PUT)
-	public String modificarProducto(@Valid @ModelAttribute("producto") Producto producto) {
-		
+	public String modificarProducto(@Valid @ModelAttribute("producto") Producto producto, HttpSession session) {
+		if(session.getAttribute("userId") != null) {
 		System.out.println("el Id a modificar es: "+producto.getId());
 		pservice.modificarProducto(producto);
-		
 		return "redirect:/producto";
+		}
+		else {
+			return "redirect:/login";
+		}
 	}
 	
 	@RequestMapping(value="/eliminar/{id}", method = RequestMethod.DELETE)
-	public String eliminar(@PathVariable("id") Long id) {
+	public String eliminar(@PathVariable("id") Long id, HttpSession session) {
+		if(session.getAttribute("userId") != null) {
 		System.out.println("Eliminar id: "+ id);
 		pservice.eliminarProducto(id);
-		
 		return "redirect:/producto";
+		}
+		else {
+			return "redirect:/login";
+		}
 	}
 	
 }
